backend/requirements.txt
fastapi
uvicorn
supabase
openai
python-dotenv
python-jose[cryptography]
pydantic